Debugging break instruction

+++ trunk/r4300/pure_interp.c
@@ -162,7 +162,8 @@ static void BREAK(){
 	int i;
 	for(i=0; i<32; i+=4)
 		printf("r%2d: %08x  r%2d: %08x  r%2d: %08x  r%2d: %08x\n",
-		       i, reg[i], i+1, reg[i+1], i+2, reg[i+2], i+3, reg[i+3]);
+		       i, (unsigned int)reg[i], i+1, (unsigned int)reg[i+1],
+		       i+2, (unsigned int)reg[i+2], i+3, (unsigned int)reg[i+3]);
 	printf("Press A to continue execution\n");
 	while(!(PAD_ButtonsHeld(0) & PAD_BUTTON_A));
 	while( (PAD_ButtonsHeld(0) & PAD_BUTTON_A));
